requirements.txt
streamlit
inference-sdk
opencv-python-headless
Pillow
numpy
gTTS
fpdf
streamlit-js-eval
requests